Create a Connection Client to a Field database (internal)

This function creates a crul::HttpClientobject for use in retrieving all documents from or querying a Field database associated with a specific project. This function is intended for internal use only.

Usage

proj_idf_client(conn, project = NULL, include = "all")

Arguments

conn

A connection object returned by connect_idaifield().

project

(deprecated) character. Name of the project-database that should be loaded.

include

Arguments: "all", "query", "changes" . Should the client use "_all_docs", "_find" or "_changes" as paths.

crul on CRAN

CouchDB API

Value

A crul::HttpClient() object.

See also

Examples

if (FALSE) { # \dontrun{
  connection <- connect_idaifield(pwd = "hallo", project = "rtest")
  client <- proj_idf_client(conn = connection)
} # }