Establish a connection to the iDAI.field / Field Desktop Client

The connection-object is used to store the connection settings needed to connect to the database of your iDAI.field / Field Desktop Client. It contains all the information that other functions such as `get_idaifield_docs()` need access the database.

Usage

connect_idaifield(
  serverip = "127.0.0.1",
  project = NULL,
  user = "R",
  pwd = "password",
  version = 3,
  ping = TRUE
)

Arguments

serverip

The IP that the user can find in the Field Clients settings as 'Your address'/'Eigene Adresse' without the port-specification (':3000')

project

character. The project you want to work with. Will be overwritten by the projectname specification in other functions. If not supplied to other function, they will use the project recorded in the connection object.

user

(optional) A user name. This should currently not be needed.

pwd

The Password as it is displayed in the Field Clients settings as 'Your password'/'Eigenes Passwort'

version

2 if iDAI.field 2 is used, 3 if you are using Field Desktop (default 3, integer).

ping

logical. default TRUE; Should the connection be tested on creation? If TRUE, also checks if the project exists.

Value

an `idaifieldR`-connection settings object (`idf_connection_settings`)

Details

If you are using Field Desktop on the same machine, you should not need the `serverip`-argument, as it defaults to the common localhost address. Likewise, the `user`-argument is currently not needed for access. `pwd` needs to be the password that is set in your Field Desktop-Client under Tools/Werkzeuge > Settings/Einstellungen: 'Your password'/'Eigenes Passwort'. If the default `serverip`-argument does not work for you, or you want to access a Client on the same network that is not running on the same machine as R, you can exchange it for the address listed above the password (without the port (':3000')). The `version`-argument does not need to be specified if you use the current version of Field Desktop (3), but will help you connect if you are using 'iDAI.field 2'. You can set the `project` that you want to work with in this function, but be aware that other functions will overwrite this setting if you supply a projectname there. `connect_idaifield()` will check if the project actually exists and throw an error if it does not.

Examples

if (FALSE) {
conn <- connect_idaifield(serverip = "127.0.0.1",
user = "R", pwd = "hallo")
}